+++ src/test/java/org/tinspin/index/test/util/PointMapCandidate.java
@@ -23,6 +23,7 @@ import org.tinspin.index.rtree.RTreeEntry;
 import org.tinspin.index.rtree.RTree;
 import org.tinspin.index.test.util.TestInstances.IDX;
 import org.tinspin.index.util.PointMapWrapper;
+import org.tinspin.index.util.Refs;
 
 import static org.tinspin.index.Index.*;
 
@@ -79,7 +80,7 @@ public class PointMapCandidate extends Candidate {
 	public void load(double[] data, int dims) {
 		this.data = data;
 		if (bulkloadSTR) {
-			RTreeEntry<Integer>[] entries = new RTreeEntry[N];
+			RTreeEntry<Integer>[] entries = Refs.newArray(RTreeEntry.class, N);
 			int pos = 0;
 			for (int i = 0; i < N; i++) {
 				double[] buf = new double[dims];