+++ CHANGELOG.md
@@ -20,6 +20,8 @@ TODO
   [#48](https://github.com/tzaeschke/tinspin-indexes/pull/48)
 - Fixed javadoc warnings
   [#49](https://github.com/tzaeschke/tinspin-indexes/pull/49)
+- Fixed linter warnings
+  [#50](https://github.com/tzaeschke/tinspin-indexes/pull/50)
 
 ## [2.1.4] - 2024-08-01